Re-added $PATH manipulation in local_env.sh
It seems that when `run.sh` is executed from Ansible using sudo, the
galaxy user's `.bashrc` is not sourced, thus `/usr/local/bin` does not
end up in the path. There are a few required binaries (such as
`scontrol`) that in the `/usr/local/bin`, thus it's been added to
`local_env.sh` to ensure it's on the $PATH when Galaxy is run.